+++ Cargo.toml
@@ -60,7 +60,6 @@ tower-http = { version = "0.6", features = ["fs", "trace", "cors"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter", "json"] }
 utoipa = { version = "5", features = ["axum_extras", "chrono", "uuid"] }
-utoipa-swagger-ui = { version = "9", features = ["axum"] }
 uuid = { version = "1.11", features = ["v4", "v7", "serde", "js"] }
 utoipa-axum = { version = "0.2" }
 pulldown-cmark = "0.12"
