+++ crates/observation-tools-server/src/ui/observation_detail.rs
@@ -26,7 +26,10 @@ pub async fn observation_detail(
   );
   let parsed_observation_id = observation_tools_shared::ObservationId::parse(&observation_id)?;
   let observation = match metadata.get_observation(parsed_observation_id).await {
-    Ok(obs) => Some(GetObservation::new(obs)),
+    Ok(obs) => {
+      let payloads = metadata.get_all_payloads(parsed_observation_id).await?;
+      Some(GetObservation::new(obs, payloads))
+    }
     Err(crate::storage::StorageError::NotFound(_)) => {
       // The user may go to the observation page before it's uploaded. Since the page
       // auto-refreshes, we do not throw an error so it will show up once it's
