+++ crates/observation-tools-server/templates/execution_detail.html
@@ -76,7 +76,7 @@
                     {% for obs in observations %}
                       <a
                         data-testid="ObservationListItemLink"
-                        href="{{ base_path }}?offset={{ offset }}&limit={{ limit }}&obs={{ obs.id }}"
+                        href="{{ base_path }}?{% if previous_page_token %}page_token={{ previous_page_token }}&{% endif %}obs={{ obs.id }}"
                         class="block px-3 py-1.5 hover:bg-neutral-800 cursor-pointer border-b border-neutral-800 last:border-b-0 {% if selected_observation and selected_observation.id == obs.id %}bg-neutral-800{% endif %}"
                       >
                         <div data-testid="ObservationListItem" class="flex items-start gap-3">
@@ -120,23 +120,22 @@
                       </a>
                     {% endfor %}
                   </div>
-                  {% if total_count %}
+                  {% if previous_page_token or has_next_page %}
                     <div class="divider"></div>
                     <p data-testid="PaginationInfo" class="mb-2 text-base-content/60 text-sm">
-                      showing {{ offset + 1 }}-{{ offset + observations|length }} of
-                      {{ total_count }} (page {{ page }})
+                      showing {{ observations|length }} observations
                     </p>
                     <div class="join">
                       <button
                         data-testid="PaginationPrev"
-                        {% if offset > 0 %}onclick="window.location.href='{{ base_path }}?offset={{ offset - limit }}&limit={{ limit }}'"{% else %}disabled{% endif %}
+                        {% if previous_page_token %}onclick="window.location.href='{{ base_path }}'"{% else %}disabled{% endif %}
                         class="join-item btn btn-sm btn-outline"
                       >
-                        &larr; prev
+                        &larr; first
                       </button>
                       <button
                         data-testid="PaginationNext"
-                        {% if has_next_page %}onclick="window.location.href='{{ base_path }}?offset={{ offset + limit }}&limit={{ limit }}'"{% else %}disabled{% endif %}
+                        {% if has_next_page %}onclick="window.location.href='{{ base_path }}?page_token={{ next_page_token }}'"{% else %}disabled{% endif %}
                         class="join-item btn btn-sm btn-outline"
                       >
                         next &rarr;
